requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.0
aiosignal==1.2.0
async-timeout==4.0.1
asynctest==0.13.0 ; python_version < '3.8'
attrs==21.2.0
bidict==0.21.4
charset-normalizer==2.0.7
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
python-engineio==4.3.0
python-socketio==5.4.1
typing-extensions==3.10.0.2 ; python_version < '3.8'
yarl==1.7.2
